HV: move common stuff from assign.c
Move common stuff, like ptdev entry and softirq, to new ptdev.c

Acked-by: Eddie Dong <eddie.dong@intel.com>
