+++ hypervisor/Makefile
@@ -141,6 +141,7 @@ C_SRCS += common/trusty_hypercall.c
 C_SRCS += common/schedule.c
 C_SRCS += common/vm_load.c
 C_SRCS += common/io_request.c
+C_SRCS += common/ptdev.c
 
 ifdef STACK_PROTECTOR
 C_SRCS += common/stack_protector.c
