+++ hypervisor/include/arch/x86/assign.h
@@ -7,66 +7,13 @@
 #ifndef ASSIGN_H
 #define ASSIGN_H
 
-enum ptdev_intr_type {
-	PTDEV_INTR_MSI,
-	PTDEV_INTR_INTX,
-	PTDEV_INTR_INV,
-};
-
-enum ptdev_vpin_source {
-	PTDEV_VPIN_IOAPIC,
-	PTDEV_VPIN_PIC,
-};
-
-/* entry per guest virt vector */
-struct ptdev_msi_info {
-	uint32_t vmsi_addr; /* virt msi_addr */
-	uint32_t vmsi_data; /* virt msi_data */
-	uint16_t vmsi_ctl; /* virt msi_ctl */
-	uint32_t pmsi_addr; /* phys msi_addr */
-	uint32_t pmsi_data; /* phys msi_data */
-	int msix;	/* 0-MSI, 1-MSIX */
-	int msix_entry_index; /* MSI: 0, MSIX: index of vector table*/
-	int virt_vector;
-	int phys_vector;
-};
-
-/* entry per guest vioapic pin */
-struct ptdev_intx_info {
-	enum ptdev_vpin_source vpin_src;
-	uint8_t virt_pin;
-	uint8_t phys_pin;
-};
-
-/* entry per each allocated irq/vector
- * it represents a pass-thru device's remapping data entry which collecting
- * information related with its vm and msi/intx mapping & interaction nodes
- * with interrupt handler and softirq.
- */
-struct ptdev_remapping_info {
-	struct vm *vm;
-	uint16_t virt_bdf;	/* PCI bus:slot.func*/
-	uint16_t phys_bdf;	/* PCI bus:slot.func*/
-	uint32_t active;	/* 1=active, 0=inactive and to free*/
-	enum ptdev_intr_type type;
-	struct dev_handler_node *node;
-	struct list_head softirq_node;
-	struct list_head entry_node;
-
-	union {
-		struct ptdev_msi_info msi;
-		struct ptdev_intx_info intx;
-	} ptdev_intr_info;
-};
+#include <ptdev.h>
 
 void ptdev_intx_ack(struct vm *vm, int virt_pin,
 		enum ptdev_vpin_source vpin_src);
 int ptdev_msix_remap(struct vm *vm, uint16_t virt_bdf,
 		struct ptdev_msi_info *info);
 int ptdev_intx_pin_remap(struct vm *vm, struct ptdev_intx_info *info);
-void ptdev_softirq(int cpu);
-void ptdev_init(void);
-void ptdev_release_all_entries(struct vm *vm);
 int ptdev_add_intx_remapping(struct vm *vm, uint16_t virt_bdf,
 	uint16_t phys_bdf, uint8_t virt_pin, uint8_t phys_pin, bool pic_pin);
 void ptdev_remove_intx_remapping(struct vm *vm, uint8_t virt_pin, bool pic_pin);
@@ -74,6 +21,5 @@ int ptdev_add_msix_remapping(struct vm *vm, uint16_t virt_bdf,
 	uint16_t phys_bdf, int vector_count);
 void ptdev_remove_msix_remapping(struct vm *vm, uint16_t virt_bdf,
 		int vector_count);
-int get_ptdev_info(char *str, int str_max);
 
 #endif /* ASSIGN_H */
